+++ packages/react/src/components/IconButton/index.tsx
@@ -1,10 +1,3 @@
-/**
- * Unfortunately, eslint does not recognize the Polymorphic component has propTypes set
- *
- * We might be able to remove this if we upgrade eslint and associated plugins
- * See: https://github.com/dequelabs/cauldron/issues/451
- */
-/* eslint-disable react/prop-types */
 import React, {
   useRef,
   forwardRef,
@@ -13,32 +6,26 @@ import React, {
   HTMLProps
 } from 'react';
 import classnames from 'classnames';
-import * as Polymorphic from '../../utils/polymorphic-type';
 import Icon, { IconType } from '../Icon';
 import Tooltip, { TooltipProps } from '../Tooltip';
 import Offscreen from '../Offscreen';
+import {
+  PolymorphicProps,
+  PolymorphicComponent
+} from '../../utils/polymorphicComponent';
+import { HTMLAttributes } from 'enzyme';
 
-export interface IconButtonOwnProps {
+export interface IconButtonProps
+  extends PolymorphicProps<React.HTMLAttributes<HTMLButtonElement>, 'button'> {
   icon: IconType;
   label: React.ReactNode;
   tooltipPlacement?: TooltipProps['placement'];
   tooltipVariant?: TooltipProps['variant'];
   tooltipPortal?: TooltipProps['portal'];
   variant?: 'primary' | 'secondary' | 'error';
+  large?: boolean;
 }
 
-type PolymorphicIconButton = Polymorphic.ForwardRefComponent<
-  'button',
-  IconButtonOwnProps
->;
-
-/**
- * Unfortunately, eslint does not recognize that this Polymorphic component has a displayName set
- *
- * We might be able to remove this if we upgrade eslint and associated plugins
- * See: https://github.com/dequelabs/cauldron/issues/451
- */
-// eslint-disable-next-line react/display-name
 const IconButton = forwardRef(
   (
     {
@@ -52,6 +39,7 @@ const IconButton = forwardRef(
       variant = 'secondary',
       disabled,
       tabIndex = 0,
+      large,
       ...other
     }: any,
     ref
@@ -81,7 +69,8 @@ const IconButton = forwardRef(
             IconButton: true,
             'IconButton--primary': variant === 'primary',
             'IconButton--secondary': variant === 'secondary',
-            'IconButton--error': variant === 'error'
+            'IconButton--error': variant === 'error',
+            'IconButton--large': large
           })}
           ref={internalRef}
           disabled={disabled}
@@ -107,7 +96,7 @@ const IconButton = forwardRef(
       </React.Fragment>
     );
   }
-) as PolymorphicIconButton;
+) as PolymorphicComponent<IconButtonProps, 'button'>;
 
 IconButton.displayName = 'IconButton';